requirements.txt
Django==1.7.8
Django-Select2==4.3.1
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.3.0
SQLAlchemy==0.9.7
South==1.0.2
Unidecode==0.04.17
Werkzeug==0.9.6
YURL==0.13
aldryn-boilerplates==0.7
aldryn-bootstrap3==0.4.5
argparse==1.2.1
boto==2.20.1
cmsplugin-filer==0.10.1
colorama==0.2.5
dj-database-url==0.3.0
django-admin-enhancer==1.0.0
django-appconf==1.0.1
django-appdata==0.1.4
django-classy-tags==0.6.1
django-cms==3.1.0
-e git+https://github.com/h3/django-colorfield.git@3e2e2a26b68483e9fe7c829fdbbb33f0202e3bf5#egg=django_colorfield-master
django-durationfield==0.5.1
django-filer==0.9.9
django-hvad==1.2.1
django-meta==0.2.1
django-meta-mixin==0.1.1
django-mptt==0.6.1
django-parler==1.4
django-polymorphic==0.7.1
django-reversion==1.8.5
django-robots==1.1
django-sekizai==0.8.1
django-sortedm2m==0.9.5
django-taggit==0.14.0
django-taggit-autosuggest==0.2.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==3.0
django-uuidfield==0.5.0
djangocms-admin-style==0.2.5
djangocms-blog==0.4.0
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.6
djangocms-link==1.6.1
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
easy-thumbnails==2.2
envdir==0.7
html5lib==0.999
httplib2==0.8
itsdangerous==0.24
lockfile==0.8
psycopg2==2.6
pytz==2015.2
requests==2.7.0
six==1.9.0
tzlocal==1.1.3
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
web.py==0.37
wsgiref==0.1.2
